requirements.txt
google-cloud-vision==0.31.0
boto3==1.7.10
python-dotenv==0.8.2
